Factor out NoteText to wrap editable/non-editable and use it in Home feed
