+++ app/assets/stylesheets/student_profile_page.scss
@@ -3,14 +3,6 @@
 .StudentProfilePage {
   font-size: 14px;
 
-  .note-text {
-    border: 1px solid transparent;
-  }
-
-  .note-text:hover {
-  	border-color: #e5e5e5;
-  }
-
   .address-book-icon {
     display: inline-block;
     height: 50px;
