+++ app/assets/stylesheets/style.scss
@@ -12,6 +12,7 @@
 @import 'components/help_bubble';
 @import 'components/flexible_roster';
 @import 'components/school_dashboard';
+@import 'components/NoteText';
 
 @import 'loader';
 @import 'roster';
